完成转向功能测试；加入手动自动的基础功能；计划执行部分写完，未实际测试

+++ plan_manage_main/src/app/include/pm_flash.h
@@ -14,7 +14,7 @@ void pm_flash_init(void);
 void flash_write(uint8_t *saddr, uint16_t nbyte);
 
 
-void flash_read(uint8_t *daddr, uint8_t nbyte);
+void flash_read(uint8_t *daddr, uint16_t nbyte);
